Reverts removing useLoadStore in order to maintain the consistency of the code
