+++ src/routes/safe/container/index.jsx
@@ -1,17 +1,13 @@
 // @flow
 import * as React from 'react'
-import { useEffect, useState } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useState } from 'react'
+import { useSelector } from 'react-redux'
 
 import Page from '~/components/layout/Page'
-import loadAddressBookFromStorage from '~/logic/addressBook/store/actions/loadAddressBookFromStorage'
-import addViewedSafe from '~/logic/currentSession/store/actions/addViewedSafe'
 import { type Token } from '~/logic/tokens/store/model/token'
 import Layout from '~/routes/safe/components/Layout'
 import { useCheckForUpdates } from '~/routes/safe/container/Hooks/useCheckForUpdates'
-import fetchLatestMasterContractVersion from '~/routes/safe/store/actions/fetchLatestMasterContractVersion'
-import fetchSafe from '~/routes/safe/store/actions/fetchSafe'
-import fetchTransactions from '~/routes/safe/store/actions/fetchTransactions'
+import { useLoadStore } from '~/routes/safe/container/Hooks/useLoadStore'
 import { safeParamAddressFromStateSelector } from '~/routes/safe/store/selectors'
 
 type Action = 'Send' | 'Receive'
@@ -27,26 +23,9 @@ const INITIAL_STATE = {
 const SafeView = () => {
   const [state, setState] = useState(INITIAL_STATE)
   const [safeLoaded, setSafeLoaded] = useState(false)
-  const dispatch = useDispatch()
-
   const safeAddress = useSelector(safeParamAddressFromStateSelector)
 
-  useEffect(() => {
-    const fetchData = () => {
-      if (safeAddress) {
-        dispatch(fetchLatestMasterContractVersion())
-          .then(() => dispatch(fetchSafe(safeAddress)))
-          .then(() => {
-            setSafeLoaded(true)
-            dispatch(loadAddressBookFromStorage())
-            return dispatch(fetchTransactions(safeAddress))
-          })
-          .then(() => dispatch(addViewedSafe(safeAddress)))
-      }
-    }
-    fetchData()
-  }, [safeAddress])
-
+  useLoadStore({ setSafeLoaded, safeAddress })
   useCheckForUpdates()
 
   const onShow = (action: Action) => () => {
